backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.21",
    "django>=6.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.11",
    "psycopg[binary]>=3.3.2",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
]
